db-bot/requirements.txt
openai>=1.0.0
qdrant-client>=1.6.0
python-dotenv
fastapi
uvicorn
supabase
pydantic
tqdm
